refactor: rename file js
